Setting Up The "Hello, World!"
I was able to print "Hello, world!' by using python coding and putting the code +print to start off the begining of the coding and then put parentheses around the quotation marks which including the text I wanted to include into the coding script. I finished the coding script off with a closing parentheses and quotation mark.